# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Urlopomat API
  version: 0.1.0
  description: >
    Poland-focused vacation planning API.
    Version 1 supports only Poland public holidays, weekend logic, workday calculations,
    leave analysis, and leave recommendations.
servers:
  - url: https://127.0.0.1:8000
...
paths:
...
  /api/v1/leave/recommendations:
    get:
      tags:
        - Leave
      operationId: getLeaveRecommendations
      summary: Get leave recommendations for selected period and leave budget
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
          example: '2026-01-01'
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
          example: '2026-12-31'
        - name: budget
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
          example: 5
        - name: strategy
          in: query
          required: false
          schema:
            type: string
            enum:
              - best_ratio
            default: best_ratio
          example: best_ratio
      responses:
        '200':
          description: Ranked leave recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveRecommendationsResponse'
              example:
                country: PL
                strategy: best_ratio
                budget: 1
                from: '2026-05-01'
                to: '2026-05-10'
                recommendations:
                  - from: '2026-05-01'
                    to: '2026-05-04'
                    total_days_off: 4
                    leave_days_required: 1
                    ratio: 4
                    holidays_count: 1
                    leave_dates:
                      - '2026-05-04'
                    holidays:
                      - code: labour_day
                        name: Swieto Pracy
                        date: '2026-05-01'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_budget:
                  value:
                    error:
                      code: missing_budget
                      message: Query parameter "budget" is required.
                invalid_budget:
                  value:
                    error:
                      code: invalid_budget
                      message: Query parameter "budget" must be an integer greater or equal to 0.
                invalid_strategy:
                  value:
                    error:
                      code: invalid_strategy
                      message: Recommendation strategy "unknown" is not supported.
                invalid_range:
                  value:
                    error:
                      code: invalid_range
                      message: Query parameter "from" must be before or equal to "to".
                range_too_large:
                  value:
                    error:
                      code: range_too_large
                      message: Date range cannot exceed 366 days.
                unsupported_year:
                  value:
                    error:
                      code: unsupported_year
                      message: Year 1999 is not supported. Minimum supported year is 2000.
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    Holiday:
      type: object
      required:
        - code
        - name
        - date
      properties:
        code:
          type: string
          example: new_year
        name:
          type: string
          example: Nowy Rok
        date:
          type: string
          format: date
          example: '2026-01-01'
...
    LeaveRecommendation:
      type: object
      required:
        - from
        - to
        - total_days_off
        - leave_days_required
        - ratio
        - holidays_count
        - leave_dates
        - holidays
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        total_days_off:
          type: integer
        leave_days_required:
          type: integer
        ratio:
          type: number
          format: float
        holidays_count:
          type: integer
        leave_dates:
          type: array
          items:
            type: string
            format: date
        holidays:
          type: array
          items:
            $ref: '#/components/schemas/Holiday'
    LeaveRecommendationsResponse:
      type: object
      required:
        - country
        - strategy
        - budget
        - from
        - to
        - recommendations
      properties:
        country:
          type: string
          example: PL
        strategy:
          type: string
          example: best_ratio
        budget:
          type: integer
          minimum: 0
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/LeaveRecommendation'